Create a CLASS that is a word guessing game.
User inputs word to be guessed by other user.
The guessing user has certain number of guesses based on the length of mystery word.
The word is converted to blank spaces corresponding to each letter and each space matching the user's guess is revealed as letter.
The game is over when the guessing user either guesses mystery word before the allotted number of turns, or does not guess word.